+++ Makefile
@@ -39,11 +39,17 @@ SRCS	= src/parser/check_element.c \
 		  src/calc/plus_minus_vector.c \
 		  src/calc/normalize.c \
 		  src/calc/distance_vector.c \
-		  src/raytracing/ray_calc.c \
 		  src/raytracing/ray_intersections.c \
 		  src/raytracing/calc_screen_vector.c \
 		  src/raytracing/calc_distance_obj.c \
-		  src/raytracing/hit_obj_calc.c
+		  src/raytracing/hit_obj_calc.c \
+		  src/raytracing/calc_camera_pos.c \
+		  src/raytracing/calc_shade.c \
+		  src/raytracing/calc_shadow.c \
+		  src/raytracing/check_light_pos.c \
+		  src/raytracing/check_nearest_obj.c \
+		  src/raytracing/clamp_utils.c \
+		  src/raytracing/color_utils.c 
 OBJS_M  = $(SRCS_M:.c=.o)
 OBJS_B  = $(SRCS_B:.c=.o)
 OBJS	= $(SRCS:.c=.o)
