+++ includes/raytracing.h
@@ -2,6 +2,7 @@
 # define RAYTRACING_H
 
 #include "./parser.h"
+#include "./ui.h"
 # define SHINENESS 16
 # define NO_HIT -99
 # define MAX_DIST 65535
@@ -52,4 +53,24 @@ double	get_nearest_inter_cy(t_obj *obj, t_xyz *h_pos, double *t, t_xyz norm_c);
 void	fill_hit_obj(t_obj *obj, t_ray c_ray, t_hit_point *h_obj);
 void	set_hit_obj(t_obj *obj, t_ray *ray, t_hit_point *h_obj, double dist);
 void	set_h_obj_cy(t_obj *obj, t_ray *ray, t_hit_point *h_obj, double *t);
+
+t_xyz	calc_cam_dir(t_xyz screen_vec, t_xyz cam_vec);
+
+int		hit_nearest_obj(t_obj *obj, t_ray *ray, t_hit_point *hit_p);
+int		hit_shadow_ray(t_obj *obj, t_ray *sh_ray, t_hit_point *hit_p);
+void	check_light_and_cam_pos(t_obj *obj, t_lit *lit, t_ray cam_ray);
+void	check_light_pos(t_obj *obj, t_env *env, t_ray cam_ray);
+
+unsigned int	clamp(double value, int min, int max);
+double			clamp_double(double value, double min, double max);
+void			clamp_xyz(t_xyz *rgb, double min, double max);
+
+t_xyz	calc_shade(t_obj *obj, t_lit *lit, t_hit_point hit_obj, t_ray cam_ray);
+
+int		calc_shadow(t_obj *obj, t_lit *lit, t_hit_point *hit_p);
+
+int	set_amb_col(t_xyz *color, t_env *env);
+void	pls_amb_color(t_obj *obj, t_env *env, t_xyz *col);
+t_xyz	pls_shade(t_obj *obj, t_lit *lit, double diff_ref, double spec_ref);
+void	color_set_to_pixel(t_meta_img *img, int x, int y, unsigned int color);
 #endif
